flask-backend/requirements.txt
# Flask and extensions
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
python-dotenv==0.19.1

# Database
SQLAlchemy==1.4.27
PyMySQL==1.0.2
mysqlclient==2.2.0
alembic==1.7.5

# Security and authentication
bcrypt==3.2.0
PyJWT==2.8.0

# API and requests
requests==2.26.0
Werkzeug==2.0.2

# NLP and ML
spacy==3.2.0
transformers==4.31.0
scikit-learn==1.0.1
pandas==1.3.4
numpy==1.21.4

# Utilities
python-dateutil==2.8.2
email-validator==2.0.0

# Additional dependencies
gunicorn==20.1.0
cryptography==36.0.0 